fixed codeClimate

+++ src/main/java/ru/mail/polis/nik27090/DAOImpl.java
@@ -110,10 +110,14 @@ public class DAOImpl implements DAO {
 
         for (int i = oldGenerate; i >= 0; i--) {
             final File file = new File(storage, i + SUFFIX);
+            final SSTable ssTable = ssTables.remove(i);
+            if (ssTable == null){
+                break;
+            }
+            ssTable.close();
             try {
-                ssTables.remove(i).close();
                 Files.delete(file.toPath());
-            } catch (NullPointerException e) {
+            } catch (NoSuchFileException e) {
                 break;
             }
         }
